+++ src/stream/filestream.cpp
@@ -25,13 +25,13 @@ namespace stream
 			fclose(f);
 	};
 
-	size_t FileStream::stream_read(void* tobuffer, size_t buffersize, size_t readamount)
+	int64_t FileStream::stream_read(void* tobuffer, int64_t buffersize, int64_t readamount)
 	{
 	
 		return fread_s(tobuffer, buffersize, 1, readamount, f);
 	}
 
-	size_t FileStream::stream_write(const void* frombuffer, size_t writesize)
+	int64_t FileStream::stream_write(const void* frombuffer, int64_t writesize)
 	{
 		return fwrite(frombuffer, 1, writesize, f);
 	}
