+++ src/stream/net_stream_audio_out.cpp
@@ -1,45 +1,3 @@
-//#include <stream/net_stream_audio_out.h>
-//
-//namespace stream
-//{
-//
-//
-//	NetStreamAudioOut::NetStreamAudioOut(wsock::udpSocket& local, wsock::addr& remote_addr, int32_t ms)
-//		: sock(local), remote(remote_addr)
-//	{
-//		memset(sendbuf, 0, sizeof(sendbuf));
-//		packet.id = 0;
-//		packet.timestamp = 0;
-//		packet.data_in_ms = ms;
-//
-//	}
-//
-//	NetStreamAudioOut::~NetStreamAudioOut(){}
-//
-//	size_t NetStreamAudioOut::stream_write(const void* frombuffer, size_t writesize)
-//	{
-//		
-//		packet.id++;
-//		packet.size = writesize;
-//		packet.timestamp = std::chrono::high_resolution_clock::now().time_since_epoch().count();
-//		packet.data = (char*)frombuffer;
-//		
-//		memcpy(sendbuf, &packet, sizeof(packet));
-//		memcpy(sendbuf + sizeof(packet), packet.data, packet.size);
-//		printf("[%u] %lli\n",packet.id,(packet.timestamp-last)/1000000);
-//		last = packet.timestamp;
-//		sock._sendto(remote, sendbuf, sizeof(packet) + packet.size, 0);
-//
-//		return 1;
-//
-//
-//	}
-//
-//	size_t NetStreamAudioOut::stream_read(void* tobuffer, size_t buffersize, size_t readamount)
-//	{
-//		return 0;
-//	};
-//}
 
 #include <stream/net_stream_audio_out.h>
 
@@ -55,7 +13,7 @@ namespace stream
 
 	NetStreamAudioOut::~NetStreamAudioOut() {}
 
-	size_t NetStreamAudioOut::stream_read(void* tobuffer, size_t buffersize, size_t readamount)
+	int64_t NetStreamAudioOut::stream_read(void* tobuffer, int64_t buffersize, int64_t readamount)
 	{
 		pack::AudioPacket ap;
 		if (jb->pop(ap) == JSUCCESS)
@@ -73,7 +31,7 @@ namespace stream
 		return 1;
 	}
 
-	size_t NetStreamAudioOut::stream_write(const void* frombuffer, size_t writesize)
+	int64_t NetStreamAudioOut::stream_write(const void* frombuffer, int64_t writesize)
 	{
 		packet_headers.packet_type = AUDIO_PACKET_TYPE;
 		packet_headers.id++;
